dmv_routes_frontend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.0.3
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
